+++ compy.files
@@ -1,14 +1,16 @@
-libcompy/types.h
-libcompy/tests.h
-libcompy/tests.c
-libcompy/result.h
-libcompy/quicksort.h
-libcompy/quicksort.c
-libcompy/main.c
-libcompy/linked_list.h
-libcompy/linked_list.c
-libcompy/huffman.h
-libcompy/huffman.c
-libcompy/file.h
-libcompy/file.c
+libcompy/libcompy.h
+libcompy/io/file.h
+libcompy/io/file.c
+libcompy/data/quicksort.h
+libcompy/data/quicksort.c
+libcompy/data/linked_list.h
+libcompy/data/linked_list.c
+libcompy/data/huffman.h
+libcompy/data/huffman.c
+libcompy/core/types.h
+libcompy/core/tests.h
+libcompy/core/tests.c
+libcompy/core/result.h
+libcompy/core/platform.h
+compy/main.c
 compy/main.c
