+++ libcompy/core/platform.h
@@ -1,0 +1,13 @@
+#ifndef __SC_PLATFORM_H
+#define __SC_PLATFORM_H
+
+
+
+
+#undef restrict
+#define restrict __restrict__
+
+
+
+
+#endif
