+++ libcompy/core/result.h
