+++ libcompy/core/tests.h
@@ -4,11 +4,6 @@
 
 
 
-#include "./types.h"
-
-
-
-
 #ifdef __cplusplus
 extern "C" {
 #endif
