+++ libcompy/core/types.h
@@ -1,0 +1,13 @@
+#ifndef __SC_TYPES_H
+#define __SC_TYPES_H
+
+
+
+
+#include <stdint.h>
+#include <stddef.h>
+
+
+
+
+#endif
