+++ libcompy/data/huffman.h
@@ -4,7 +4,8 @@
 
 
 
-#include "./types.h"
+#include "../core/platform.h"
+#include "../core/result.h"
 
 #include "./linked_list.h"
