+++ libcompy/data/linked_list.c
