+++ libcompy/data/linked_list.h
@@ -4,7 +4,10 @@
 
 
 
-#include "./types.h"
+#include "../core/platform.h"
+#include "../core/result.h"
+
+#include "./quicksort.h"
