+++ libcompy/libcompy.h
@@ -1,0 +1,15 @@
+#ifndef __LIBCOMPY_LIBCOMPY_H
+#define __LIBCOMPY_LIBCOMPY_H
+
+
+
+
+#include "./data/huffman.h"
+#include "./data/quicksort.h"
+
+#include "./io/file.h"
+
+
+
+
+#endif
